refactor: move FileSelector components and BreadcrumbSegment to shared locations
Move FileSelector components from BrowsePage/FileSelector/ to FileSelector/
and BreadcrumbSegment from BrowsePage/ to widgets/ for better reusability.
Update import paths in consumers.
