+++ frontend/src/components/ui/widgets/BreadcrumbSegment.tsx
